+++ scripts/check-publish.js
@@ -18,22 +18,38 @@
 const path = require('path');
 const chalk = require('chalk').default;
 const cp = require('child_process');
+const fs = require('fs');
 
-let code = 0;
-const workspaces = JSON.parse(cp.execSync('yarn --silent workspaces info').toString());
-for (const name in workspaces) {
-    const workspace = workspaces[name];
-    const location = path.resolve(process.cwd(), workspace.location);
-    const packagePath = path.resolve(location, 'package.json');
-    const pck = require(packagePath);
-    if (!pck.private) {
-        const pckName = `${pck.name}@${pck.version}`;
-        if (cp.execSync(`npm view ${pckName} version --json`).toString().trim()) {
-            console.info(`${pckName}: published`);
-        } else {
-            console.error(`(${chalk.red('ERR')}) ${pckName}: ${chalk.red('NOT')} published`);
-            code = 1;
+checkPublish().catch(error => {
+    console.error(error);
+    process.exitCode = 1;
+});
+
+async function checkPublish() {
+    const workspaces = JSON.parse(cp.execSync('yarn --silent workspaces info').toString());
+    await Promise.all(Object.values(workspaces).map(async workspace => {
+        const packagePath = path.resolve(workspace.location, 'package.json');
+        const pck = JSON.parse(await fs.promises.readFile(packagePath, 'utf8'));
+        if (!pck.private) {
+            const pckName = `${pck.name}@${pck.version}`;
+            const npmViewOutput = await new Promise(
+                resolve => cp.exec(`npm view ${pckName} version --json`,
+                    (error, stdout, stderr) => {
+                        if (error) {
+                            console.error(error);
+                            resolve('');
+                        } else {
+                            resolve(stdout.trim());
+                        }
+                    }
+                )
+            );
+            if (npmViewOutput) {
+                console.info(`${pckName}: published`);
+            } else {
+                console.error(`(${chalk.red('ERR')}) ${pckName}: ${chalk.red('NOT')} published`);
+                process.exitCode = 1;
+            }
         }
-    }
+    }));
 }
-process.exit(code);